Remove Platform checks helper and use selenium checks helper
Move ci rake task to rakefile to almost completely clear up top level incorrectly namespaced files
