+++ Rakefile
@@ -33,7 +33,6 @@ require 'rake_tasks/selenium_rake/ie_code_generator'
 require 'rake_tasks/selenium_rake/java_formatter'
 require 'rake_tasks/selenium_rake/cpp_formatter'
 require 'rake_tasks/selenium_rake/type_definitions_generator'
-require 'rake_tasks/ci'
 
 # Our modifications to the Rake library
 require 'rake_tasks/rake/task'
@@ -453,7 +452,6 @@ end
 
 namespace :ci do
   task :upload_to_sauce do
-
     upload_path = ENV['UPLOAD_PATH']
     username = ENV['SAUCE_USERNAME']
     apikey = ENV['SAUCE_APIKEY']
@@ -492,8 +490,27 @@ namespace :ci do
   end
 end
 
+# Required for above ci:upload_to_sauce rake only
+require 'uri'
+require 'net/http'
+require 'digest/md5'
+require 'json'
+require 'pathname'
+
+def net_http
+  http_proxy = ENV['http_proxy'] || ENV['HTTP_PROXY']
+  if http_proxy
+    http_proxy = "http://#{http_proxy}" unless http_proxy.start_with?('http://')
+    proxy_uri = URI.parse(http_proxy)
+
+    Net::HTTP::Proxy(proxy_uri.host, proxy_uri.port)
+  else
+    Net::HTTP
+  end
+end
+
 at_exit do
-  if File.exist?(".git") && !Platform.windows?
+  if File.exist?(".git") && !SeleniumRake::Checks.windows?
     system "sh", ".git-fixfiles"
   end
 end
